Write methods in user and plant model
Method in user model to create a list of users with unwatered plants.
Method in plant model to make a list of plant ids that are unwatered.
Write tasks in scheluder for heroku's scheluder to do every 24 hrs.
Check datatbase for water_events with watered:false and then text those users that their plants need watering.